jarvis/requirements.txt
fastapi
uvicorn[standard]
google-adk
google-genai
python-dotenv
